🚚 move [#37] : coupon 디렉토리 이동
