+++ docker-compose.yml
@@ -1,7 +1,7 @@
 services:
   coupon-service:
     container_name: coupon-service
-    image: chokyuseong/coupon-service:scheduler
+    image: chokyuseong/coupon-service:latest
     env_file:
       - .env
     build:
