+++ src/main/java/com/sparta/limited/coupon_service/application/dto/request/CouponCreateRequest.java
@@ -1,0 +1,33 @@
+package com.sparta.limited.coupon_service.application.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class CouponCreateRequest {
+
+    private final String name;
+    private final Integer discountRate;
+    private final Long quantity;
+
+    private CouponCreateRequest(
+        String name,
+        Integer discountRate,
+        Long quantity
+    ) {
+        this.name = name;
+        this.discountRate = discountRate;
+        this.quantity = quantity;
+    }
+
+    public static CouponCreateRequest of(
+        String name,
+        Integer discountRate,
+        Long quantity
+    ) {
+        return new CouponCreateRequest(
+            name,
+            discountRate,
+            quantity
+        );
+    }
+}
