+++ src/main/java/com/sparta/limited/coupon_service/application/dto/response/CouponReadOneResponse.java
@@ -1,0 +1,46 @@
+package com.sparta.limited.coupon_service.application.dto.response;
+
+import com.sparta.limited.coupon_service.domain.model.CouponStatus;
+import java.util.UUID;
+import lombok.Getter;
+
+@Getter
+public class CouponReadOneResponse {
+
+    private final UUID couponId;
+    private final String name;
+    private final Integer discountRate;
+    private final CouponStatus status;
+    private final Long quantity;
+
+    private CouponReadOneResponse(
+        UUID couponId,
+        String name,
+        Integer discountRate,
+        CouponStatus status,
+        Long quantity
+    ) {
+        this.couponId = couponId;
+        this.name = name;
+        this.discountRate = discountRate;
+        this.status = status;
+        this.quantity = quantity;
+    }
+
+    public static CouponReadOneResponse of(
+        UUID couponId,
+        String name,
+        Integer discountRate,
+        CouponStatus status,
+        Long quantity
+    ) {
+        return new CouponReadOneResponse(
+            couponId,
+            name,
+            discountRate,
+            status,
+            quantity
+        );
+    }
+
+}
