--- src/main/java/com/sparta/limited/coupon_service/coupon/domain/repository/CouponRepository.java
@@ -1,11 +1,0 @@
-package com.sparta.limited.coupon_service.coupon.domain.repository;
-
-import com.sparta.limited.coupon_service.coupon.domain.model.Coupon;
-import java.util.UUID;
-
-public interface CouponRepository {
-
-    void save(Coupon coupon);
-
-    Coupon findById(UUID couponId);
-}
