--- src/main/java/com/sparta/limited/coupon_service/coupon/infrastructure/persistence/JpaCouponRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.limited.coupon_service.coupon.infrastructure.persistence;
-
-import com.sparta.limited.coupon_service.coupon.domain.model.Coupon;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaCouponRepository extends JpaRepository<Coupon, UUID> {
-
-}
