+++ src/main/java/com/sparta/limited/coupon_service/domain/repository/CouponRepository.java
@@ -1,0 +1,11 @@
+package com.sparta.limited.coupon_service.domain.repository;
+
+import com.sparta.limited.coupon_service.domain.model.Coupon;
+import java.util.UUID;
+
+public interface CouponRepository {
+
+    void save(Coupon coupon);
+
+    Coupon findById(UUID couponId);
+}
