+++ src/main/java/com/sparta/limited/coupon_service/infrastructure/redis/RedisScriptKey.java
@@ -1,0 +1,18 @@
+package com.sparta.limited.coupon_service.infrastructure.redis;
+
+import lombok.Getter;
+
+@Getter
+public enum RedisScriptKey {
+    QUANTITY_AND_DUPLICATE("lua/quantity-duplicate.lua"),
+    ;
+
+    private final String luaScripName;
+
+    RedisScriptKey(
+        String luaScripName
+    ) {
+        this.luaScripName = luaScripName;
+    }
+
+}
